Internal Competition
The University of Iowa Business Analytics Case Competition is a contest that consists of 16 universities across the nation, and 4 people for each university. My master's program department decided to choose these 4 individuals by having an internal contest amongst our cohort. This was the work I submitted to be determined as the final 4. The goal was to predict the classification of an individuals obesity level given data on specified variables. In my work, I applied machine learning techniques in my logistic, random forest, and neural network model. I then chose the right classification based off the similarities between the predictions of the respective model.