Case Competition
This was the work that we prepared for our case competition. We were given survey data as well as HR and Operational performance metrics of 16 of their manufacturing plants. Our job was to analyze this data, identify key factors that effect our performance metrics, and offer suggestions on how we can use this information to improve future performance metrics for each manufacturing plant.